upload video

+++ src/middlewares/multer.js
@@ -4,6 +4,7 @@ import fs from 'fs';
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
+    console.log('🚀 ~ file: multer.js ~ line 7 ~ file', file);
     const fieldname = file.fieldname;
     if (fieldname === 'avatar') {
       cb(null, path.join(__dirname, '../uploads/avatar'));
@@ -14,22 +15,27 @@ const storage = multer.diskStorage({
     }
   },
   filename: function (req, file, cb) {
+    console.log('🚀 ~ file: multer.js ~ line 18 ~ file', file);
     const user = req.user?.id;
-    const typeImage = file.originalname.split('.')[1];
-    fs.readdirSync(path.join(__dirname, `../uploads/${file.fieldname}`))
-      .filter((file) => {
-        const f = file.split('.');
-        return f[0];
-      })
-      .forEach((existFile) => {
-        if (existFile.split('.')[0] === `${user + file.fieldname}`) {
-          fs.unlinkSync(
-            path.join(__dirname, `../uploads/${file.fieldname}/${existFile}`),
-          );
-        }
-      });
-    const uniqueSuffix = '.' + typeImage;
-    cb(null, user + file.fieldname + uniqueSuffix);
+    if (file.fieldname === 'avatar') {
+      const typeImage = file.originalname.split('.')[1];
+      fs.readdirSync(path.join(__dirname, `../uploads/${file.fieldname}`))
+        .filter((file) => {
+          const f = file.split('.');
+          return f[0];
+        })
+        .forEach((existFile) => {
+          if (existFile.split('.')[0] === `${user + file.fieldname}`) {
+            fs.unlinkSync(
+              path.join(__dirname, `../uploads/${file.fieldname}/${existFile}`),
+            );
+          }
+        });
+      const uniqueSuffix = '.' + typeImage;
+      cb(null, user + file.fieldname + uniqueSuffix);
+    } else {
+      cb(null, user + file.fieldname + file.originalname);
+    }
   },
 });
